:truck: Rename(admin/component): 수정 관리자 컴포넌트 폴더 구조
