+++ src/app/(main)/admin/_components/slider-manage.tsx
